Remove `Data.Map.Diff.Strict.Internal` module.
The internal module has become obsolete, so we move all its contents
into its parent module.
